refactor : api 로직 분리 (리팩토링)
